+++ src/info.c
@@ -38,7 +38,7 @@ void logger(int level, int priority, const char *format, ...) {
 }
 
 char *strip_weight(char *netstr) {
-	int len = strlen(netstr);
+	size_t len = strlen(netstr);
 
 	if(len >= 3 && !strcmp(netstr + len - 3, "#10")) {
 		netstr[len - 3] = 0;
