+++ src/openssl/digest.h
@@ -27,7 +27,7 @@ struct digest {
 	const EVP_MD *digest;
 	HMAC_CTX *hmac_ctx;
 	EVP_MD_CTX *md_ctx;
-	int maclength;
+	size_t maclength;
 };
 
 #endif
