+++ src/openssl/rsagen.c
@@ -81,7 +81,7 @@ rsa_t *rsa_generate(size_t bits, unsigned long exponent) {
 	BN_set_word(bn_e, exponent);
 	BN_GENCB_set(cb, indicator, NULL);
 
-	int result = RSA_generate_key_ex(rsa, bits, bn_e, cb);
+	int result = RSA_generate_key_ex(rsa, (int) bits, bn_e, cb);
 
 	BN_GENCB_free(cb);
 	BN_free(bn_e);
